Added tests for existing behaviour
The existing gem functionality expected to be run within Rails, e.g. assuming
that `cattr_accessor`/`mattr_accessor` is available. To keep the library useable
_as purely Rack middleware_, I've emulated the `cattr_accessor`-like behaviour
and removed the implicit use of ActiveSupport methods.

I've used Minitest for the test suite to keep the footprint relatively light and
the tests fast. Also added Coveralls to generate code coverage for the test suite.